xwm: Split module
